feat(ui): Add workflow execution logs viewer in workflows-sidebar
This adds a new view in the workflows sidebar body that shows different
logs from the session instead of workflow settings. The default
behaviour for now is to show these logs after the workflow is ran. This
might change in the future depending on user preferences. In that case,
the revert is easy and you can still access these in the chevron down
button in the new button group footer.